First Half of MlShareMapState in TS
