+++ src/components/MlShareMapState/MlShareMapState.tsx
@@ -4,6 +4,7 @@ import PropTypes from 'prop-types';
 import MapContext from '../../contexts/MapContext';
 import { v4 as uuidv4 } from 'uuid';
 import useMapState from '../../hooks/useMapState';
+import MapLibreGlWrapper from '../MapLibreMap/lib/MapLibreGlWrapper';
 
 const getCurrentUrlParameters = () => {
 	const currentParams = Object.fromEntries(new URLSearchParams(window.location.search));
@@ -23,12 +24,19 @@ const initialUrlParams = getCurrentUrlParameters();
  * @component
  */
 
-const MlShareMapState = (props) => {
+export interface MlShareMapStateProps {
+	//
+	mapId?: string;
+	idPrefix?: string;
+	active?: boolean;
+}
+
+const MlShareMapState = (props: MlShareMapStateProps) => {
 	// Use a useRef hook to reference the layer object to be able to access it later inside useEffect hooks
 	const mapContext = useContext(MapContext);
 	const initializedRef = useRef(false);
-	const mapRef = useRef(undefined);
-	const [map, setMap] = useState(undefined);
+	const mapRef = useRef<MapLibreGlWrapper | undefined>();
+	const [map, setMap] = useState<MapLibreGlWrapper | undefined>(undefined);
 	const layersFromUrlParamsRef = useRef({});
 	const componentId = useRef((props.idPrefix ? props.idPrefix : 'MlShareMapState-') + uuidv4());
 	const mapState = useMapState({
@@ -65,22 +73,23 @@ const MlShareMapState = (props) => {
 		const mapLayers = [];
 		for (const x in mapState.layers) {
 			mapLayers.push({
-				id: mapState.layers[x].id,
-				type: mapState.layers[x].type,
-				visible: mapState.layers[x].visible,
+				id: mapState.layers[x]?.id,
+				type: mapState.layers[x]?.type,
+				visible: mapState.layers[x]?.visible,
 			});
 		}
 		refreshMapState();
-		let urlParams = new URLSearchParams({
+		const urlParams = new URLSearchParams({
 			...getCurrentUrlParameters(),
 			...mapStateRef.current,
 			layers: JSON.stringify(mapLayers),
 		});
-		JSON.parse(Object.fromEntries(urlParams).layers).forEach((el) => {
+		JSON.parse(Object.fromEntries(urlParams).layers).forEach((el: { id: number }) => {
+			// is iD a number?
 			layersFromUrlParamsRef.current[el.id] = false;
 		});
 
-		let currentParams = new URLSearchParams(window.location.search);
+		const currentParams = new URLSearchParams(window.location.search);
 		checkRestorationStates(mapState.layers);
 		if (urlParams.toString() !== currentParams.toString()) {
 			window.history.pushState(
@@ -92,7 +101,7 @@ const MlShareMapState = (props) => {
 	}, [mapState.layers, props.active]);
 
 	useEffect(() => {
-		let _componentId = componentId.current;
+		const _componentId = componentId.current;
 
 		mapStateRef.current = getCurrentUrlParameters();
 
@@ -104,7 +113,7 @@ const MlShareMapState = (props) => {
 			// check for the existence of map.style before calling getLayer or getSource
 
 			if (mapRef.current) {
-				mapRef.current.cleanup(_componentId);
+				mapRef.current.cleanup(_componentId); //Vergleich z.39
 				mapRef.current = undefined;
 			}
 			initializedRef.current = false;
@@ -120,12 +129,12 @@ const MlShareMapState = (props) => {
 	useEffect(() => {
 		if (!mapRef.current) return;
 
-		let _refreshUrlParameters = refreshUrlParameters;
+		const _refreshUrlParameters = refreshUrlParameters;
 
 		mapRef.current.on('moveend', _refreshUrlParameters, componentId.current);
 
 		return () => {
-			mapRef.current?.off('moveend', _refreshUrlParameters);
+			mapRef.current?.wrapper.off('moveend', _refreshUrlParameters); // welches "off" aus MapLibreGLWrapper ist das richtige?
 		};
 	}, [refreshUrlParameters, map]);
 
@@ -137,6 +146,7 @@ const MlShareMapState = (props) => {
 		mapRef.current = mapContext.getMap(props.mapId);
 		setMap(mapRef.current);
 		if (mapStateRef.current.lat && mapStateRef.current.lng) {
+			// Keine Ahnung (?)
 			restoreViewportState();
 		}
 	}, [mapContext.mapIds, mapContext, props.mapId, props.active]);
